+++ src/helpers/mod.rs
@@ -4,8 +4,11 @@ pub use echo_x_request_id::echo_x_request_id;
 mod with_shared_data;
 pub use with_shared_data::with_shared_data;
 
-mod error_with_message;
-pub use error_with_message::ErrorWithMessage;
+mod error_with_static_message;
+pub use error_with_static_message::ErrorWithStaticMessage;
+
+mod error_with_dynamic_message;
+pub use error_with_dynamic_message::ErrorWithDynamicMessage;
 
 mod unit_conversion;
 pub use unit_conversion::convert_metres_to_degrees;